FIX.
Database now ensures unique SpecialPassword name and shortcut

+++ src/db/Database.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Comparator;
 import java.util.HashMap;
+import java.util.Optional;
 import java.util.TreeMap;
 import java.util.Vector;
 import java.util.stream.Collectors;
@@ -113,9 +114,31 @@ public class Database
         }
     }
 
+    /*
+     * Method evaluates if password fits in database.
+     * It shall be called whenever new entry is added to database.
+     * Current rules:
+     * 1) password name must be unique
+     * 2) password shortcut must be empty or unique
+     * NOTE: Order of checks matters for replaceEntry();
+     */
+    private void validatePassword(SpecialPassword entry) throws Exceptions
+    {
+        if (!entry.getShortcut().isEmpty())
+        {
+            Optional<SpecialPassword> optSp =
+                db.keySet().stream().filter(sp -> sp.getShortcut().equals(entry.getShortcut())).findFirst();
+            if (optSp.isPresent())
+                throw new Exceptions(XC.PASSWORD_SHORTCUT_ALREADY_IN_USE).setText(optSp.get().getName());
+            // TODO: some smarter way to pass info through exceptions (maybe pass copy of whole SpecialPassword?)
+        }
+
+        if (db.containsKey(entry)) throw new Exceptions(XC.PASSWORD_NAME_ALREADY_EXISTS);
+    }
+
     public void addEntry(SpecialPassword entry) throws Exceptions
     {
-        if (db.containsKey(entry)) throw new Exceptions(XC.PASSWORD_NAME_ALREADY_EXISTS);
+        validatePassword(entry);
 
         db.put(entry, rsa.encrypt(Utilities.objectToBytes(entry.getMap())));
 
@@ -123,8 +146,9 @@ public class Database
         sync();
     }
 
-    public void deleteEntry(SpecialPassword entry)
+    public void deleteEntry(SpecialPassword entry) throws Exceptions
     {
+        if (!db.containsKey(entry)) throw new Exceptions(XC.NO_PASSWORD_FOUND);
         db.remove(entry);
 
         status = Status.DESYNCHRONIZED;
@@ -133,10 +157,26 @@ public class Database
 
     public void replaceEntry(SpecialPassword newEntry, SpecialPassword oldEntry) throws Exceptions
     {
-        // This check should be before anything else to avoid loosing oldEntry
+        // All checks should be before anything else to avoid loosing oldEntry
         // (either replace is completed fully, or not at all)
-        if (!oldEntry.getName().equals(newEntry.getName()) && db.containsKey(newEntry))
-            throw new Exceptions(XC.PASSWORD_NAME_ALREADY_EXISTS);
+
+        if (!db.containsKey(oldEntry)) throw new Exceptions(XC.NO_PASSWORD_FOUND);
+
+        try
+        {
+            validatePassword(newEntry);
+        }
+        catch (Exceptions e)
+        {
+            // Password may have existing name, if it will substitute password with this name
+            if (e.getCode() == XC.PASSWORD_NAME_ALREADY_EXISTS && oldEntry.getName().equals(newEntry.getName()))
+                Logger.printDebug("Replacing same SpecialPassword");        // Not an error
+            // Password may have conflicting shortcut with itself
+            else if (e.getCode() == XC.PASSWORD_SHORTCUT_ALREADY_IN_USE && e.getText().equals(newEntry.getName()))
+                Logger.printDebug("Didn't change shortcut for password");   // Not an error
+            else
+                throw e;        // Error
+        }
         db.remove(oldEntry);
         db.put(newEntry, rsa.encrypt(Utilities.objectToBytes(newEntry.getMap())));
